Final commit

+++ Micromouse/src/Ultra.h
@@ -5,9 +5,9 @@ void setupUltra();
 
 byte loopUltra();
 
-int getDistance0();
-int getDistance1();
-int getDistance2();
+float getDistance0();
+float getDistance1();
+float getDistance2();
